[BQ] one small emulator update, one small source config update (Recidiviz/recidiviz-data#37698)
## Description of the change

this PR makes two unrelated changes that are required for
https://github.com/Recidiviz/recidiviz-data/pull/37672

1. updates the emulator to source psuedocolumns from `SourceTableConfig`
to allow for time-based partition
2. allows for `build_source_table_repository_for_collected_schemata` to
accept `None` as the project id. this is done as some source tables
(i.e. billing tables) are only in prod. since we are building our view
graph for all projects, we also need to be able to build our source
table graph for all projects.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for https://github.com/Recidiviz/recidiviz-data/issues/37278

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 993d34e8ff8e069e50049bc8aa87517ea43b38d8