[Ingest] Update IngestViewContentsContext with stricter types and no abstract class (Recidiviz/recidiviz-data#37782)
## Description of the change

This PR is in a way a design discussion... but first here's what happens
in this PR.

1. Collapses `IngestViewContentsContext` and
`IngestViewContentsContextImpl` into just `IngestViewContentsContext`.
This was the only implementation. The documentation on the parent class
was "here's an abstraction" and the child class was "this is the
implementation". The current docstring is not much better.
2. Removes `FakeIngestViewContentsContext`, which had _extra
functionality_! Notably, it had the ability to pull environment
variables generally while parsing. This is why the production class had
`Union[str, bool]` return typing despite only returning a `bool`. In
production we are only using the context to check against the
environment for certain conditions. The test for this extra
functionality made use of
`recidiviz/tests/ingest/direct/ingest_mappings/fixtures/ingest_view_file_parser/manifests/string_datetime_env_property.yaml`

So, my understanding is production use of `IngestViewContentsContext`
gates parsing functionality on project environments. Most importantly we
have the `should_launch` manifest node. However, we retain the ability
to have _conditional parsing based on the project_ in the ingest view.
I'm not sure if we do this in production currently, but I do not think
we should. An example of this is in
`recidiviz/tests/ingest/direct/ingest_mappings/fixtures/ingest_view_file_parser/manifests/boolean_condition_env_property.yaml`

I don't see a reason to parse ingest view results differently across
different projects, and would prefer if this was only for
`should_launch`. Should we go that direction, we can stop passing
`IngestViewContentsContext` into all of the `build_from_row` methods and
maybe even tighten up what `should_launch` is and can be, as it is
currently fairly general.

Regardless of the above, this PR brings a consistent use of
`IngestViewContentsContext` from production and testing (we actually
didn't use the Fake version of this in integration tests). Based on our
discussion here I will update the docstrings before merging.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports efforts for Recidiviz/recidiviz-data#36159

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2cb8d3edc89e79819b6610f8f78799e4bcbcd25f