[Auth] Remove or set `blocked_on` during roster ingest (Recidiviz/recidiviz-data#37231)
## Description of the change

This PR introduces new behavior during roster ingest to identify users
who should/should not be blocked and when. Users who exist in Roster but
are not included in the newest ingest are flagged with a block date one
week in the future. Any user information from Roster that doesn't exist
in UserOverride is copied over, then the user is deleted from Roster.
However, if someone is flagged to be blocked by mistake (i.e., they are
mistakenly left out of the roster ingest) and then they correctly appear
in the next ingest, their block date will be removed.

I added test cases for 4 situations:

1. A user has an upcoming block but appears in the ingested file. Their
block date will be set to null.
2. A user has an existing block but appears in the ingested file. Their
block date will remain unchanged.
3. A user does not exist in the ingested file and already exists in
UserOverride. Only null columns in their UserOverride record are updated
with their Roster information, and a block date one week in the future
is added.
4. A user does not exist in the ingested file and also does not exist in
UserOverride. A record is created in UserOverride for them with their
Roster information, and a block date one week in the future is added.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#36604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 14ffff2d10eb07a1fde625275670d1471ff0f673