[Justice Counts] update breakdown descriptions to match CSG audit (Recidiviz/recidiviz-data#37776)
## Description of the change

This PR updates the breakdown definitions for several metric
descriptions, as requested by CSG.

| Sector | Metric Name | Dimension Key | Revised Dimension Description |

|---------------------|-----------------------------|-------------------------------|----------------------------------------------------------------------------------------------------------------|
| Courts and Pretrial | Sentences Imposed | Biological Sex Male | The
number of cases in which the court imposed a sentence as a result of
criminal conviction with a defendant whose biological sex is male. |
| Courts and Pretrial | Sentences Imposed | Biological Sex Female | The
number of cases in which the court imposed a sentence as a result of
criminal conviction with a defendant whose biological sex is female. |
| Courts and Pretrial | Sentences Imposed | Biological Sex Unknown | The
number of cases in which the court imposed a sentence as a result of
criminal conviction with a defendant whose biological sex is not known.
|
| Defense | Open Cases | Case Severity Types Other | The number of
people with open criminal cases that are not felony or misdemeanor
cases. |
| Jails | Post-adjudication Admissions | Offense Types Person Offenses |
The number of post-adjudication admission events in which the most
serious offense was for an offense against a person. |
| Jails | Post-adjudication Admissions | Offense Types Property Offenses
| The number of post-adjudication admission events in which the most
serious offense was for a property offense. |
| Jails | Post-adjudication Daily Population | Biological Sex Male | A
single day count of the number of people incarcerated in the agency's
jurisdiction who have been adjudicated whose biological sex is male. |
| Jails | Post-adjudication Daily Population | Biological Sex Female | A
single day count of the number of people incarcerated in the agency's
jurisdiction who have been adjudicated whose biological sex is female. |
| Jails | Post-adjudication Daily Population | Biological Sex Unknown |
A single day count of the number of people incarcerated in the agency's
jurisdiction who have been adjudicated whose biological sex is not
known. |
| Jails | Post-adjudication Releases | Post-adjudication Releases Due to
Escape or AWOL Status | The number of post-adjudication release events
due to escape from custody or assessment as AWOL for more than 30 days.
|
| Jails | Pre-adjudication Admissions | Offense Types Person Offenses |
A single day count of the number of people incarcerated pre-adjudication
whose most serious charge was an offense against a person. |
| Jails | Pre-adjudication Admissions | Offense Types Property Offenses
| A single day count of the number of people incarcerated
pre-adjudication whose most serious charge was a property offense. |
| Jails | Pre-adjudication Admissions | Offense Types Drug Offenses | A
single day count of the number of people incarcerated pre-adjudication
whose most serious charge was a drug offense. |
| Jails | Pre-adjudication Admissions | Offense Types Public Order
Offenses | A single day count of the number of people incarcerated
pre-adjudication whose most serious charge was a public order offense. |
| Jails | Pre-adjudication Admissions | Offense Types Other Offenses | A
single day count of the number of people incarcerated pre-adjudication
whose most serious charge was not a person, property, drug, or public
order offense. |
| Jails | Pre-adjudication Admissions | Offense Types Unknown Offenses |
A single day count of the number of people incarcerated pre-adjudication
whose most serious charge type is not known. |
| Jails | Pre-adjudication Daily Population | Biological Sex Male | A
single day count of the number of people incarcerated in the agency's
jurisdiction who have not been adjudicated whose biological sex is male.
|
| Jails | Pre-adjudication Daily Population | Biological Sex Female | A
single day count of the number of people incarcerated in the agency's
jurisdiction who have not been adjudicated whose biological sex is
female. |
| Jails | Pre-adjudication Daily Population | Biological Sex Unknown | A
single day count of the number of people incarcerated in the agency's
jurisdiction who have not been adjudicated whose biological sex is not
known. |
| Jails | Total Daily Population | Offense Types Drug Offenses | A
single day count of the total number of people incarcerated under the
agency's jurisdiction in which the most serious charge was for a drug
offense. |
| Prisons | Releases | Release Types Other Releases from Prison | The
number of release events from the agency's prison jurisdiction that are
not releases to probation supervision, to parole supervision, to other
community supervision, with no additional correctional control, or due
to death. |
| Prosecution | Cases Referred | Referred Case Severity Misdemeanor
Cases Referred | The number of criminal cases referred to the office in
which the leading charge was for a misdemeanor offense. |
es

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 46a66f7639506c04df9951af8eff95dae847ef4f