[Justice Counts] Send over metric details for agency in admin panel, even if it is not a superagency (Recidiviz/recidiviz-data#37779)
This PR addresses a bug discovered by Ilya and Mahmoud, where users
attempting to copy metric settings to child agencies would encounter an
issue if they hadn't saved their change after designating an agency as a
superagency. Specifically, the metric settings wouldn't show up in the
dropdown for the user to select and copy over.

**Reproduction of Bug**
In the attached video:

https://github.com/user-attachments/assets/fe3455aa-ff0b-4967-8dc6-37caa46be8f5

A user attempts to designate an agency as a superagency.
Without saving the change, the user proceeds to copy over metric
settings to child agencies.
The metric settings fail to appear in the dropdown for selection,
causing the process to break.

**Resolution**
This PR ensures that metric data is sent for an agency regardless of
whether it is currently marked as a superagency. This allows users to
copy over metric settings to child agencies without needing to save the
superagency change beforehand.

**Testing**
I tested this change on the staging environment by deploying the updates
to my playtesting branch and replicating the steps from the bug. The fix
worked as expected.

## Related issues

Closes Recidiviz/recidiviz-data#37700

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37a9e66f6b926a68eadf84bff4392d4171afaa33