[US_UT] Update Housing Status Periods to use ofndr_addr_hist (Recidiviz/recidiviz-data#37835)
## Description of the change
Previously we were using ofndr_addr_arch. Turns out that stands for
`archive` (I think) and it's only updated until 2022. The new table is
ofndr_addr_hist, but this data has its own issues. Namely, rather than
having the address encoded in the table itself like it was in
ofndr_addr_arch, the specific address appears to have been split out
into a different table with an addr_id. However, we don't have this
table. This means that we're left only knowing the type of address
(addr_typ_cd), not the specific address. Also, to reiterate, we don't
actually know what the addr_typ_cd's map too -- we lack that mapping
table as well -- so all my mappings are educated guesses based off the
free-text comments I see in the table.

Sandbox:
https://console.cloud.google.com/dataflow/jobs/us-west1/2025-01-28_20_47_28-11303851128425188386?project=recidiviz-staging

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5a9878f6ff76e62b37ad138d558444a91f048639