[US_NE] Sentence length and sentence group length (Recidiviz/recidiviz-data#37557)
## Description of the change

Ingesting sentence length and sentence group length for Nebraska. The
`Offense` table in NE contains sentence and charge information, with the
county attached to the Offense table to indicate where they were
sentenced for each specific offense. Each `inmateNumber` will only have
one record in the `AggregateSentence` table because all of the
individual offenses in the `Offense` table get rolled up into one and
each `inmateNumber` only lasts for one full cycle of incarceration ->
supervision -> liberty. That is why we are only able to join Offense and
AggregateSentence on inmateNumber and it should be accurate. This is
also why we treat `inmateNumber` and `pkAggregateSentenceId` as a
sentence group and attach the `pkOffenseId` for individual sentences.
Still, we expect all the major dates (besides the min/max days) to be
the same since they edit the `Offense` table dates to equal the
AggregateSentence dates.

Succesful pipeline run
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2025-01-30_07_54_45-14569795481483554309?authuser=0&organizationId=448885369991&project=recidiviz-staging).

I've purposely included only one test for view because we expect a few
iterations of correcting sentence information with Nebraska. Their data
is incredibly messy and incomplete, with much information not being
entered into their system. They often refer to presentencing reports and
note files not entered anywhere we would be able to access.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f3a8bfba915f18efd467bbeb73f9713202baf191