[US_AR] Allow some E1 locations in IP view (Recidiviz/recidiviz-data#37892)
## Description of the change

Previously, E1 ("County Jail/Sheriff") locations were not considered
"facilities" in the AR IP view, meaning that spans of time in these
locations were not ingested as incarceration periods. This is still true
in most cases, but with one exception we weren't accounting for prior to
this change. Most of the time that an E1 location is involved in a
movement in EXTERNALMOVEMENT, it's in a purely jurisdictional capacity;
for instance, if a movement is reported by an E1 location with a
movement type of "out to court", "out to furlough", "out to field
supv.", etc., it means that the person is under the jurisdiction of the
location, but is not detained in any physical location. HOWEVER, when a
movement has a movement code of 38 (used for arrested supervision
violators) into an E1 location, it means that the location (usually a
sheriff's department) **is** physically detaining the person, usually
for a brief period of time before further processing. Likewise,
movements with the code 42 (a paired code with 38, used for released
jail detainees) indicate releases from detention in an E1 location.

Therefore, E1 has been added to the list of allowed location types in
the facility_locs CTE. In order to keep jurisdictional-only movements
from being treated as facility movements, we append JURISDICTIONAL_ONLY
to all E1 location IDs in EXTERNALMOVEMENT if they don't have a movement
code of 38 or 42, so that when we check the location ID against the list
of allowed facility-type location IDs, the JURISDICTIONAL_ONLY IDs won't
have a match. No post-processing is needed for these location IDs, since
any period in a JURISDICTIONAL_ONLY location will ultimately be filtered
out.

This change should address _some_ of the SPs with an "incarcerated"
status that don't have corresponding IPs. Most of the time, though, SPs
with an "incarcerated" status won't have connected data in
EXTERNALMOVEMENT since their detention is managed by the courts, not the
DOC. Further work may be necessary to start inferring incarceration
periods using the supervision data.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#37763

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3f0de9ad37f633a39ff614d8832be78e9e974a1c